Update Field-level completeness chart and helper functions
