+++ ui-tests/completeness-page.ui.ts
@@ -2,6 +2,7 @@ import { test, expect } from "@playwright/test";
 import dotenv from "dotenv";
 import path from "path";
 import { VIEWPORT, TIMEOUTS, DEBUG, FILE_PATHS } from './constants';
+import { fieldLevelCompletenessTest } from "./helpers";
 
 dotenv.config({ path: path.resolve(__dirname, "../.env") });
 
@@ -75,80 +76,6 @@ test.describe("Completeness page", async () => {
     await expect(tooltip).toContainText(expectedValue);
   }
 
-  async function testFieldLevelHoverText({
-    page,
-    graphTitle,
-    barIndex,
-    expectedLabel,
-    expectedValue,
-  }) {
-    // Locate the bar based on the graph title and bar index
-    const selectedBar = await page
-      .getByText(graphTitle)
-      .locator("..")
-      .locator("..")
-      .locator('.highcharts-root > .highcharts-series-group > .highcharts-series > path')
-      .nth(barIndex)
-
-    // Hover over the selected bar
-    await expect(selectedBar).toBeVisible();
-    // Doesn't seem to work under any circumstances with Highcharts
-    /* if (barIndex > 5) {
-      // Scroll the scrollbar a bit down
-      await page.locator('.highcharts-scrollbar-thumb').hover({ force: true });
-      await page.mouse.down();
-      await page.mouse.move(page.mouse._x, page.mouse._y + 100);
-      await page.mouse.up();
-    } */
-    // NOTE: force: true is needed because highcharts intercepts the normal hover event
-    await selectedBar.hover({ force: true });
-
-    // Locate the tooltip relative to the graph title
-    const tooltip = await page
-      .getByText(graphTitle)
-      .locator("..")
-      .locator("..")
-      .locator(".highcharts-tooltip")
-      .nth(0);
-
-    // Verify the tooltip text
-    await expect(tooltip).toContainText(expectedLabel);
-    await expect(tooltip).toContainText(expectedValue);
-  }
-
-  async function testFieldLevel(page, testCases, programName) {
-
-    const fieldLevelGraph = await page
-      .getByText("Field Level")
-      .locator("..")
-      .locator("..")
-      .last();
-
-    await expect(fieldLevelGraph).toHaveScreenshot(`FieldLevel-${programName}.png`, {
-      maxDiffPixelRatio: 0.05,
-    });
-    // Sort the percentages
-    testCases.sort((a, b) => a.label.localeCompare(b.label));
-    testCases.sort((a, b) => a.pct - b.pct);
-    testCases.forEach((datum, index) => {
-      datum.value = `${datum.pct}%`;
-      datum.barIndex = index;
-    });
-    for (const {label, value, barIndex } of testCases) {
-      // DEBUG: Currently there is no way to scroll this particular graph
-      // So, we'll do only one page of results
-      if (barIndex > 10) {
-        break;
-      }
-      await testFieldLevelHoverText({
-        page,
-        graphTitle: "Field Level",
-        barIndex,
-        expectedLabel: label,
-        expectedValue: value,
-      });
-    }
-  }
   //#endregion
 
   //#region Page Overview
@@ -236,75 +163,8 @@ test.describe("Completeness page", async () => {
   //#endregion
 
   //#region Field Level
-  test('Field-level Completeness', async ({ request }) => {
-    async function getEndpoint(page, endpoint) {
-      const { cookies } = await page.context().storageState();
-      const sessionCookie = cookies.find(cookie => cookie.name === "session_id");
-      let headers = {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${sessionCookie.value}`,
-      };
-      const url = `${BASE_URL}/${endpoint}`;
-      return fetch(url, { headers });
-    }
-
-    // Wait for the page to finish loading the field level completeness
-    await expect(page.getByText("Field Level")
-      .locator("..")
-      .locator(".."))
-      .toHaveText(/.+Radiations.+/i);
-
-    // Query the discovery/programs endpoint
-    const response = await getEndpoint(page, "query/discovery/programs");
-    await response.json().then(async (data) => {
-      let lastButtonText = /All programs/;
-      const allCases = {};
-      for (const program of data.programs) {
-        const programButton = await page.getByText(lastButtonText).first();
-        // 1: Switch the display to being this particular program
-        await programButton.click();
-        lastButtonText = new RegExp(`.+ ${program.program_id}`, "i");
-        await page.getByRole('option', { name: lastButtonText }).click();
-
-        const completenessData = program.metadata.required_but_missing;
-        const categories = Object.keys(completenessData);
-        const testCases = categories.map((category) => {
-          return Object.keys(completenessData[category]).map((key) => {
-            const label = `${category}/${key}`;
-            const pct = Math.round((1 - (completenessData[category][key]['missing'] / completenessData[category][key]['total'])) * 100);
-
-            // Fill out the allCases for the final round
-            if (label in allCases) {
-              allCases[label]['missing'] += completenessData[category][key]['missing'];
-              allCases[label]['total'] += completenessData[category][key]['total'];
-            } else {
-              allCases[label] = {
-                'missing': completenessData[category][key]['missing'],
-                'total': completenessData[category][key]['total']
-              }
-            }
-            return { label, pct, value: 'NA', barIndex: 0 };
-          });
-        }).flat(1);
-
-        await testFieldLevel(page, testCases, program.program_id);
-      }
-
-      // Do one final round for the "all programs" option
-      await page.getByText(lastButtonText).first().click();
-      await page.getByRole('option', { name: "All programs" }).click();
-      await page.waitForTimeout(1000);
-
-      const allCasesList = Object.keys(allCases).map((thisCase) => {
-        return ({
-          label: thisCase,
-          pct: Math.round((1 - (allCases[thisCase].missing / allCases[thisCase].total)) * 100),
-          value: 'NA',
-          barIndex: 0
-        });
-      });
-      await testFieldLevel(page, allCasesList, 'allprograms');
-    });
+  test('Field-level Completeness', async () => {
+    await fieldLevelCompletenessTest(page, BASE_URL);
   });
   //#endregion
