+++ ui-tests/helpers.ts
@@ -11,6 +11,7 @@ export async function login(page, username, password) {
     await page.getByRole("button", { name: "Sign In" }).click();
 }
 
+// Charts
 export async function testBarGraphHoverText({
     page,
     graphTitle,
@@ -69,6 +70,7 @@ export async function testStackedBarGraphHoverText({
     await expect(tooltip).toContainText(expectedValue);
 }
 
+// Search Page Componenets 
 export async function fillCheckbox(page, checkboxSelector, value) {
     // Fill checkbox in dropdowns
     try {
@@ -228,3 +230,158 @@ export const selectDrugs = async (page, drugs) => {
         await fillCheckbox(page, '#checkboxes-tags-drug_name', drug);
     }
 };
+
+// Endpoint call
+export async function getEndpoint(page, endpoint, BASE_URL) {
+    const { cookies } = await page.context().storageState();
+    const sessionCookie = cookies.find(cookie => cookie.name === "session_id");
+    const token = sessionCookie.value.split('|')[1]; 
+    let headers = {
+      "Content-Type": "application/json",
+      "Authorization": `Bearer ${token}`,
+    };
+    const url = `${BASE_URL}${endpoint}`;
+    return fetch(url, { 
+      method: "GET",
+      headers 
+    });
+}
+
+// Field Level Completeness 
+export async function testFieldLevelHoverText({
+    page,
+    graphTitle,
+    barIndex,
+    expectedLabel,
+    expectedValue,
+  }) {
+    // Locate the bar based on the graph title and bar index
+    const selectedBar = await page
+      .getByText(graphTitle)
+      .locator("..")
+      .locator("..")
+      .locator('.highcharts-root > .highcharts-series-group > .highcharts-series > path')
+      .nth(barIndex)
+
+    // Hover over the selected bar
+    await expect(selectedBar).toBeVisible();
+    // Doesn't seem to work under any circumstances with Highcharts
+    /* if (barIndex > 5) {
+      // Scroll the scrollbar a bit down
+      await page.locator('.highcharts-scrollbar-thumb').hover({ force: true });
+      await page.mouse.down();
+      await page.mouse.move(page.mouse._x, page.mouse._y + 100);
+      await page.mouse.up();
+    } */
+    // NOTE: force: true is needed because highcharts intercepts the normal hover event
+    await selectedBar.hover({ force: true });
+
+    // Locate the tooltip relative to the graph title
+    const tooltip = await page
+      .getByText(graphTitle)
+      .locator("..")
+      .locator("..")
+      .locator(".highcharts-tooltip")
+      .nth(0);
+
+    // Verify the tooltip text
+    await expect(tooltip).toContainText(expectedLabel);
+    await expect(tooltip).toContainText(expectedValue);
+  }
+
+export async function testFieldLevel(page, testCases, programName) {
+
+    const fieldLevelGraph = await page
+      .getByText("Field Level")
+      .locator("..")
+      .locator("..")
+      .last();
+
+    await expect(fieldLevelGraph).toHaveScreenshot(`FieldLevel-${programName}.png`, {
+      maxDiffPixelRatio: 0.05,
+    });
+    // Sort the percentages
+    testCases.sort((a, b) => a.label.localeCompare(b.label));
+    testCases.sort((a, b) => a.pct - b.pct);
+    testCases.forEach((datum, index) => {
+      datum.value = `${datum.pct}%`;
+      datum.barIndex = index;
+    });
+    for (const {label, value, barIndex } of testCases) {
+      // DEBUG: Currently there is no way to scroll this particular graph
+      // So, we'll do only one page of results
+      if (barIndex > 10) {
+        break;
+      }
+      await testFieldLevelHoverText({
+        page,
+        graphTitle: "Field Level",
+        barIndex,
+        expectedLabel: label,
+        expectedValue: value,
+      });
+    }
+  }
+
+export async function fieldLevelCompletenessTest(page, BASE_URL) {
+    // Wait for the page to finish loading the field level completeness component
+    await expect(page.getByText("Field Level")
+      .locator("..")
+      .locator(".."))
+      .toHaveText(/.+Radiations.+/i);
+  
+    // Use your existing getEndpoint helper to fetch the data
+    const response = await getEndpoint(page, "query/discovery/programs", BASE_URL);
+    const data = await response.json();
+  
+    let lastButtonText = /All programs/;
+    const allCases = {};
+  
+    for (const program of data.programs) {
+      const programButton = await page.getByText(lastButtonText).first();
+      // Switch display to this particular program
+      await programButton.click();
+  
+      lastButtonText = new RegExp(`.+ ${program.program_id}`, "i");
+      await page.getByRole('option', { name: lastButtonText }).click();
+  
+      const completenessData = program.metadata.required_but_missing;
+      const categories = Object.keys(completenessData);
+  
+      const testCases = categories.flatMap(category => 
+        Object.keys(completenessData[category]).map(key => {
+          const label = `${category}/${key}`;
+          const pct = Math.round((1 - (completenessData[category][key].missing / completenessData[category][key].total)) * 100);
+  
+          // Accumulate for allCases
+          if (allCases[label]) {
+            allCases[label].missing += completenessData[category][key].missing;
+            allCases[label].total += completenessData[category][key].total;
+          } else {
+            allCases[label] = {
+              missing: completenessData[category][key].missing,
+              total: completenessData[category][key].total,
+            };
+          }
+  
+          return { label, pct, value: 'NA', barIndex: 0 };
+        })
+      );
+  
+      await testFieldLevel(page, testCases, program.program_id);
+    }
+  
+    // Final round for "all programs"
+    await page.getByText(lastButtonText).first().click();
+    await page.getByRole('option', { name: "All programs" }).click();
+    await page.waitForTimeout(1000);
+  
+    const allCasesList = Object.keys(allCases).map(label => ({
+      label,
+      pct: Math.round((1 - (allCases[label].missing / allCases[label].total)) * 100),
+      value: 'NA',
+      barIndex: 0,
+    }));
+  
+    await testFieldLevel(page, allCasesList, 'allprograms');
+  }
